Moving Affiliate type and enum from ExhibitForm to entity
