+++ lib/lambda/_lib/dao/dao-entity.ts
@@ -4,14 +4,13 @@ import { DAOEntity } from './dao';
 import { convertFromApiObject } from './db-object-builder';
 import { Builder, getUpdateCommandBuilderInstance } from './db-update-builder';
 import { Entity, EntityFields, YN } from './entity';
-import { DynamoDbConstruct } from '../../../DynamoDb';
 
 export const ENTITY_WAITING_ROOM:string = '__UNASSIGNED__';
 
-const dbclient = new DynamoDBClient({ region: process.env.REGION });
-const TableName = process.env[DynamoDbConstruct.DYNAMODB_ENTITY_TABLE_NAME] || '';
-
 export function EntityCrud(entityInfo:Entity, _dryRun:boolean=false): DAOEntity {
+
+  const dbclient = new DynamoDBClient({ region: process.env.REGION });
+  const TableName = process.env.DYNAMODB_ENTITY_TABLE_NAME || '';
 
   let { entity_id, entity_name, create_timestamp, update_timestamp, active=YN.Yes } = entityInfo;
