+++ lib/lambda/functions/consenting-person/ExhibitEmail.ts
@@ -1,9 +1,9 @@
 import { SESv2Client, SendEmailCommand, SendEmailCommandInput, SendEmailResponse } from "@aws-sdk/client-sesv2";
-import { AffiliateTypes, ExhibitData, ExhibitForm, IExhibitForm } from "../../_lib/pdf/ExhibitForm";
+import { v4 as uuidv4 } from 'uuid';
+import { AffiliateTypes, Entity, YN } from "../../_lib/dao/entity";
+import { ExhibitData, ExhibitForm, IExhibitForm } from "../../_lib/pdf/ExhibitForm";
 import { ExhibitFormFull } from "../../_lib/pdf/ExhibitFormFull";
 import { ExhibitFormSingle } from "../../_lib/pdf/ExhibitFormSingle";
-import { v4 as uuidv4 } from 'uuid';
-import { Entity, YN } from "../../_lib/dao/entity";
 
 
 export const enum FormTypes { FULL = 'full', SINGLE = 'single' };
